chore(refactor): create the node package
