+++ cmd/masa-node/main.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/sirupsen/logrus"
 
-	masa "github.com/masa-finance/masa-oracle/pkg"
+	"github.com/masa-finance/masa-oracle/node"
 	"github.com/masa-finance/masa-oracle/pkg/api"
 	"github.com/masa-finance/masa-oracle/pkg/config"
 	"github.com/masa-finance/masa-oracle/pkg/db"
@@ -74,22 +74,22 @@ func main() {
 	isValidator := cfg.Validator
 
 	// Create a new OracleNode
-	node, err := masa.NewOracleNode(ctx, config.EnableStaked)
+	masaNode, err := node.NewOracleNode(ctx, config.EnableStaked)
 	if err != nil {
 		logrus.Fatal(err)
 	}
-	err = node.Start()
+	err = masaNode.Start()
 	if err != nil {
 		logrus.Fatal(err)
 	}
 
-	node.NodeTracker.GetAllNodeData()
+	masaNode.NodeTracker.GetAllNodeData()
 	if cfg.TwitterScraper && cfg.DiscordScraper && cfg.WebScraper {
 		logrus.Warn("[+] Node is set as all types of scrapers. This may not be intended behavior.")
 	}
 
 	if cfg.AllowedPeer {
-		cfg.AllowedPeerId = node.Host.ID().String()
+		cfg.AllowedPeerId = masaNode.Host.ID().String()
 		cfg.AllowedPeerPublicKey = keyManager.HexPubKey
 		logrus.Infof("[+] Allowed peer with ID: %s and PubKey: %s", cfg.AllowedPeerId, cfg.AllowedPeerPublicKey)
 	} else {
@@ -97,16 +97,16 @@ func main() {
 	}
 
 	// Init cache resolver
-	db.InitResolverCache(node, keyManager)
+	db.InitResolverCache(masaNode, keyManager)
 
 	// Subscribe and if actor start monitoring actor workers
 	// considering all that matters is if the node is staked
 	// and other peers can do work we only need to check this here
 	// if this peer can or cannot scrape or write that is checked in other places
-	if node.IsStaked {
-		node.Host.SetStreamHandler(config.ProtocolWithVersion(config.WorkerProtocol), workers.GetWorkHandlerManager().HandleWorkerStream)
-		go masa.SubscribeToBlocks(ctx, node)
-		go node.NodeTracker.ClearExpiredWorkerTimeouts()
+	if masaNode.IsStaked {
+		masaNode.Host.SetStreamHandler(config.ProtocolWithVersion(config.WorkerProtocol), workers.GetWorkHandlerManager().HandleWorkerStream)
+		go node.SubscribeToBlocks(ctx, masaNode)
+		go masaNode.NodeTracker.ClearExpiredWorkerTimeouts()
 	}
 
 	// Listen for SIGINT (CTRL+C)
@@ -116,7 +116,7 @@ func main() {
 	// Cancel the context when SIGINT is received
 	go func() {
 		<-c
-		nodeData := node.NodeTracker.GetNodeData(node.Host.ID().String())
+		nodeData := masaNode.NodeTracker.GetNodeData(masaNode.Host.ID().String())
 		if nodeData != nil {
 			nodeData.Left()
 		}
@@ -130,7 +130,7 @@ func main() {
 		}
 	}()
 
-	router := api.SetupRoutes(node)
+	router := api.SetupRoutes(masaNode)
 	go func() {
 		err = router.Run()
 		if err != nil {
@@ -139,7 +139,7 @@ func main() {
 	}()
 
 	// Get the multiaddress and IP address of the node
-	multiAddr := node.GetMultiAddrs()                          // Get the multiaddress
+	multiAddr := masaNode.GetMultiAddrs()                      // Get the multiaddress
 	ipAddr, err := multiAddr.ValueForProtocol(multiaddr.P_IP4) // Get the IP address
 	// Display the welcome message with the multiaddress and IP address
 	config.DisplayWelcomeMessage(multiAddr.String(), ipAddr, keyManager.EthAddress, isStaked, isValidator, cfg.TwitterScraper, cfg.TelegramScraper, cfg.DiscordScraper, cfg.WebScraper, versioning.ApplicationVersion, versioning.ProtocolVersion)
