+++ node/oracle_node.go
@@ -1,15 +1,9 @@
-package masa
+package node
 
 import (
-	"bytes"
 	"context"
-	"encoding/base64"
-	"encoding/json"
 	"fmt"
-	"os"
-	"reflect"
 	"strings"
-	"sync"
 	"time"
 
 	ethereumCrypto "github.com/ethereum/go-ethereum/crypto"
@@ -30,15 +24,12 @@ import (
 	"github.com/multiformats/go-multiaddr"
 	"github.com/sirupsen/logrus"
 
-	shell "github.com/ipfs/go-ipfs-api"
-	pubsub "github.com/libp2p/go-libp2p-pubsub"
-
 	"github.com/masa-finance/masa-oracle/internal/versioning"
 	"github.com/masa-finance/masa-oracle/pkg/chain"
 	"github.com/masa-finance/masa-oracle/pkg/config"
 	"github.com/masa-finance/masa-oracle/pkg/masacrypto"
 	myNetwork "github.com/masa-finance/masa-oracle/pkg/network"
-	pubsub2 "github.com/masa-finance/masa-oracle/pkg/pubsub"
+	"github.com/masa-finance/masa-oracle/pkg/pubsub"
 )
 
 type OracleNode struct {
@@ -47,10 +38,9 @@ type OracleNode struct {
 	priorityAddrs     multiaddr.Multiaddr
 	multiAddrs        []multiaddr.Multiaddr
 	DHT               *dht.IpfsDHT
-	Context           context.Context
 	PeerChan          chan myNetwork.PeerEvent
-	NodeTracker       *pubsub2.NodeEventTracker
-	PubSubManager     *pubsub2.Manager
+	NodeTracker       *pubsub.NodeEventTracker
+	PubSubManager     *pubsub.Manager
 	Signature         string
 	IsStaked          bool
 	IsValidator       bool
@@ -60,10 +50,11 @@ type OracleNode struct {
 	IsWebScraper      bool
 	IsLlmServer       bool
 	StartTime         time.Time
-	WorkerTracker     *pubsub2.WorkerEventTracker
+	WorkerTracker     *pubsub.WorkerEventTracker
 	BlockTracker      *BlockEventTracker
 	Blockchain        *chain.Chain
 	options           config.AppOption
+	Context           context.Context
 }
 
 // GetMultiAddrs returns the priority multiaddr for this node.
@@ -147,7 +138,7 @@ func NewOracleNode(ctx context.Context, opts ...config.Option) (*OracleNode, err
 		return nil, err
 	}
 
-	subscriptionManager, err := pubsub2.NewPubSubManager(ctx, hst)
+	subscriptionManager, err := pubsub.NewPubSubManager(ctx, hst)
 	if err != nil {
 		return nil, err
 	}
@@ -156,9 +147,9 @@ func NewOracleNode(ctx context.Context, opts ...config.Option) (*OracleNode, err
 		Host:              hst,
 		Protocol:          config.ProtocolWithVersion(config.OracleProtocol),
 		multiAddrs:        myNetwork.GetMultiAddressesForHostQuiet(hst),
-		Context:           ctx,
 		PeerChan:          make(chan myNetwork.PeerEvent),
-		NodeTracker:       pubsub2.NewNodeEventTracker(versioning.ProtocolVersion, cfg.Environment, hst.ID().String()),
+		NodeTracker:       pubsub.NewNodeEventTracker(versioning.ProtocolVersion, cfg.Environment, hst.ID().String()),
+		Context:           ctx,
 		PubSubManager:     subscriptionManager,
 		IsStaked:          o.IsStaked,
 		IsValidator:       cfg.Validator,
@@ -196,9 +187,14 @@ func (node *OracleNode) Start() (err error) {
 	node.Host.SetStreamHandler(node.Protocol, node.handleStream)
 	node.Host.SetStreamHandler(config.ProtocolWithVersion(config.NodeDataSyncProtocol), node.ReceiveNodeData)
 
+	for pid, n := range node.options.ProtocolHandlers {
+		node.Host.SetStreamHandler(pid, n)
+	}
+
 	if node.IsStaked {
 		node.Host.SetStreamHandler(config.ProtocolWithVersion(config.NodeGossipTopic), node.GossipNodeData)
 	}
+
 	node.Host.Network().Notify(node.NodeTracker)
 
 	go node.ListenToNodeTracker()
@@ -211,7 +207,7 @@ func (node *OracleNode) Start() (err error) {
 		publicKeyHex = masacrypto.KeyManagerInstance().EthAddress
 	}
 
-	myNodeData := pubsub2.GetSelfNodeData(node.Host, node.IsStaked, node.priorityAddrs, publicKeyHex)
+	myNodeData := pubsub.GetSelfNodeData(node.Host, node.IsStaked, node.priorityAddrs, publicKeyHex)
 
 	bootstrapNodes, err := myNetwork.GetBootNodesMultiAddress(append(config.GetInstance().Bootnodes, node.options.Bootnodes...))
 	if err != nil {
@@ -226,6 +222,10 @@ func (node *OracleNode) Start() (err error) {
 	err = myNetwork.WithMDNS(node.Host, config.Rendezvous, node.PeerChan)
 	if err != nil {
 		return err
+	}
+
+	for _, p := range node.options.Services {
+		go p(node.Context, node.Host)
 	}
 
 	go myNetwork.Discover(node.Context, node.Host, node.DHT, node.Protocol)
@@ -359,200 +359,4 @@ func (node *OracleNode) LogActiveTopics() {
 	} else {
 		logrus.Info("[-] No active topics.")
 	}
-}
-
-// Blockchain Implementation
-var (
-	blocksCh = make(chan *pubsub.Message)
-)
-
-type BlockData struct {
-	Block            uint64      `json:"block"`
-	InputData        interface{} `json:"input_data"`
-	TransactionHash  string      `json:"transaction_hash"`
-	PreviousHash     string      `json:"previous_hash"`
-	TransactionNonce int         `json:"nonce"`
-}
-
-type Blocks struct {
-	BlockData []BlockData `json:"blocks"`
-}
-
-type BlockEvents struct{}
-
-type BlockEventTracker struct {
-	BlockEvents []BlockEvents
-	BlockTopic  *pubsub.Topic
-	mu          sync.Mutex
-}
-
-// HandleMessage processes incoming pubsub messages containing block events.
-// It unmarshals the message data into a slice of BlockEvents and appends them
-// to the tracker's BlockEvents slice.
-func (b *BlockEventTracker) HandleMessage(m *pubsub.Message) {
-	var blockEvents any
-
-	// Try to decode as base64 first
-	decodedData, err := base64.StdEncoding.DecodeString(string(m.Data))
-	if err == nil {
-		m.Data = decodedData
-	}
-
-	// Try to unmarshal as JSON
-	err = json.Unmarshal(m.Data, &blockEvents)
-	if err != nil {
-		// If JSON unmarshal fails, try to interpret as string
-		blockEvents = string(m.Data)
-	}
-
-	b.mu.Lock()
-	defer b.mu.Unlock()
-
-	switch v := blockEvents.(type) {
-	case []BlockEvents:
-		b.BlockEvents = append(b.BlockEvents, v...)
-	case BlockEvents:
-		b.BlockEvents = append(b.BlockEvents, v)
-	case map[string]interface{}:
-		// Convert map to BlockEvents struct
-		newBlockEvent := BlockEvents{}
-		// You might need to add logic here to properly convert the map to BlockEvents
-		b.BlockEvents = append(b.BlockEvents, newBlockEvent)
-	case []interface{}:
-		// Convert each item in the slice to BlockEvents
-		for _, item := range v {
-			if be, ok := item.(BlockEvents); ok {
-				b.BlockEvents = append(b.BlockEvents, be)
-			}
-		}
-	case string:
-		// Handle string data
-		newBlockEvent := BlockEvents{}
-		// You might need to add logic here to properly convert the string to BlockEvents
-		b.BlockEvents = append(b.BlockEvents, newBlockEvent)
-	default:
-		logrus.Warnf("[-] Unexpected data type in message: %v", reflect.TypeOf(v))
-	}
-
-	blocksCh <- m
-}
-
-func updateBlocks(ctx context.Context, node *OracleNode) error {
-
-	var existingBlocks Blocks
-	blocks := chain.GetBlockchain(node.Blockchain)
-
-	for _, block := range blocks {
-		var inputData interface{}
-		err := json.Unmarshal(block.Data, &inputData)
-		if err != nil {
-			inputData = string(block.Data) // Fallback to string if unmarshal fails
-		}
-
-		blockData := BlockData{
-			Block:            block.Block,
-			InputData:        base64.StdEncoding.EncodeToString([]byte(fmt.Sprintf("%v", inputData))),
-			TransactionHash:  fmt.Sprintf("%x", block.Hash),
-			PreviousHash:     fmt.Sprintf("%x", block.Link),
-			TransactionNonce: int(block.Nonce),
-		}
-		existingBlocks.BlockData = append(existingBlocks.BlockData, blockData)
-	}
-	jsonData, err := json.Marshal(existingBlocks)
-	if err != nil {
-		return err
-	}
-
-	err = node.DHT.PutValue(ctx, "/db/blocks", jsonData)
-	if err != nil {
-		logrus.Warningf("[-] Unable to store block on DHT: %v", err)
-	}
-
-	if os.Getenv("IPFS_URL") != "" {
-
-		infuraURL := fmt.Sprintf("https://%s:%s@%s", os.Getenv("PID"), os.Getenv("PS"), os.Getenv("IPFS_URL"))
-		sh := shell.NewShell(infuraURL)
-
-		jsonBytes, err := json.Marshal(jsonData)
-		if err != nil {
-			logrus.Errorf("[-] Error marshalling JSON: %s", err)
-		}
-
-		reader := bytes.NewReader(jsonBytes)
-
-		hash, err := sh.AddWithOpts(reader, true, true)
-		if err != nil {
-			logrus.Errorf("[-] Error persisting to IPFS: %s", err)
-		} else {
-			logrus.Printf("[+] Ledger persisted with IPFS hash: https://dwn.infura-ipfs.io/ipfs/%s\n", hash)
-			_ = node.DHT.PutValue(ctx, "/db/ipfs", []byte(fmt.Sprintf("https://dwn.infura-ipfs.io/ipfs/%s", hash)))
-
-		}
-	}
-
-	return nil
-}
-
-func SubscribeToBlocks(ctx context.Context, node *OracleNode) {
-	if !node.IsValidator {
-		return
-	}
-
-	go func() {
-		err := node.Blockchain.Init()
-		if err != nil {
-			logrus.Error(err)
-		}
-	}()
-
-	updateTicker := time.NewTicker(time.Second * 60)
-	defer updateTicker.Stop()
-
-	for {
-		select {
-		case block, ok := <-blocksCh:
-			if !ok {
-				logrus.Error("[-] Block channel closed")
-				return
-			}
-			if err := processBlock(node, block); err != nil {
-				logrus.Errorf("[-] Error processing block: %v", err)
-				// Consider adding a retry mechanism or circuit breaker here
-			}
-
-		case <-updateTicker.C:
-			logrus.Info("[+] blockchain tick")
-			if err := updateBlocks(ctx, node); err != nil {
-				logrus.Errorf("[-] Error updating blocks: %v", err)
-				// Consider adding a retry mechanism or circuit breaker here
-			}
-
-		case <-ctx.Done():
-			logrus.Info("[+] Context cancelled, stopping block subscription")
-			return
-		}
-	}
-}
-
-func processBlock(node *OracleNode, block *pubsub.Message) error {
-	blocks := chain.GetBlockchain(node.Blockchain)
-	for _, b := range blocks {
-		if string(b.Data) == string(block.Data) {
-			return nil // Block already exists
-		}
-	}
-
-	if err := node.Blockchain.AddBlock(block.Data); err != nil {
-		return fmt.Errorf("[-] failed to add block: %w", err)
-	}
-
-	if node.Blockchain.LastHash != nil {
-		b, err := node.Blockchain.GetBlock(node.Blockchain.LastHash)
-		if err != nil {
-			return fmt.Errorf("[-] failed to get last block: %w", err)
-		}
-		b.Print()
-	}
-
-	return nil
 }
