+++ pkg/config/options.go
@@ -1,10 +1,20 @@
 package config
 
+import (
+	"context"
+
+	"github.com/libp2p/go-libp2p/core/host"
+	"github.com/libp2p/go-libp2p/core/network"
+	"github.com/libp2p/go-libp2p/core/protocol"
+)
+
 type AppOption struct {
-	DisableCLIParse bool
-	IsStaked        bool
-	Bootnodes       []string
-	RandomIdentity  bool
+	DisableCLIParse  bool
+	IsStaked         bool
+	Bootnodes        []string
+	RandomIdentity   bool
+	Services         []func(ctx context.Context, node host.Host)
+	ProtocolHandlers map[protocol.ID]network.StreamHandler
 }
 
 type Option func(*AppOption)
@@ -30,5 +40,20 @@ func (a *AppOption) Apply(opts ...Option) {
 func WithBootNodes(bootnodes ...string) Option {
 	return func(o *AppOption) {
 		o.Bootnodes = append(o.Bootnodes, bootnodes...)
+	}
+}
+
+func WithService(plugins ...func(ctx context.Context, node host.Host)) Option {
+	return func(o *AppOption) {
+		o.Services = append(o.Services, plugins...)
+	}
+}
+
+func WithProtocolHandler(pid protocol.ID, n network.StreamHandler) Option {
+	return func(o *AppOption) {
+		if o.ProtocolHandlers == nil {
+			o.ProtocolHandlers = make(map[protocol.ID]network.StreamHandler)
+		}
+		o.ProtocolHandlers[pid] = n
 	}
 }
