+++ pkg/tests/integration/tracker_test.go
@@ -5,7 +5,7 @@ import (
 	"context"
 	"fmt"
 
-	. "github.com/masa-finance/masa-oracle/pkg"
+	. "github.com/masa-finance/masa-oracle/node"
 	"github.com/masa-finance/masa-oracle/pkg/config"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
