+++ pkg/workers/worker_selection.go
@@ -6,14 +6,14 @@ import (
 	"github.com/libp2p/go-libp2p/core/peer"
 	"github.com/sirupsen/logrus"
 
-	masa "github.com/masa-finance/masa-oracle/pkg"
+	"github.com/masa-finance/masa-oracle/node"
 	"github.com/masa-finance/masa-oracle/pkg/pubsub"
 	data_types "github.com/masa-finance/masa-oracle/pkg/workers/types"
 )
 
 // GetEligibleWorkers Uses the new NodeTracker method to get the eligible workers for a given message type
 // I'm leaving this returning an array so that we can easily increase the number of workers in the future
-func GetEligibleWorkers(node *masa.OracleNode, category pubsub.WorkerCategory) ([]data_types.Worker, *data_types.Worker) {
+func GetEligibleWorkers(node *node.OracleNode, category pubsub.WorkerCategory) ([]data_types.Worker, *data_types.Worker) {
 	workers := make([]data_types.Worker, 0)
 	nodes := node.NodeTracker.GetEligibleWorkerNodes(category)
 	var localWorker *data_types.Worker
